backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
fastembed
pydantic
requests
duckduckgo-search
